Now BloatWeak also detect the vulnerabilities of non-bloated dependencies

+++ PyCD/bloatWeak.py
@@ -21,11 +21,15 @@ def printHelpSection():
 
 
 #print on the shell the list of all bloated dependencies, the ones affected by vulnerabilities are highlighted with 'VULNERABLE!!!' message on the same line
-def getShellOutputJ(savepath):
+def getShellOutput(savepath):
     print("Bloated dependencies for your project:")
 
     for d in dependencies:
-        print(d.printDependency())
+        if(d.bloated):
+            if(d.vulnerable):
+                print(d.printDependency()+" VULNERABLE!!!")
+            else:
+                print(d.printDependency())
 
 
 def getSafetyDBOut(savepath):
@@ -33,19 +37,18 @@ def getSafetyDBOut(savepath):
     
     jcontents=json.load(file3)
     
-    cols=["bloated dep","version","filepath","cve","affected versions","advisory"]
+    cols=["dep","version","filepath","bloated","cve","affected versions","advisory"]
     
     deps=[]
     
     for d in dependencies:
         lowD=d.name.lower()#normalize the name of dependency as the ones in the db
         dv=d.version
-        print(lowD)
         try:
             depVul=jcontents[lowD]#extract from the db, the section related this dependency
             for v in depVul:
                 if(versionUtils.checkVersions(dv,v)):
-                    d.vulnerable="VULNERABLE!!!"
+                    d.vulnerable=True
                     cve=v['cve']
                     vers=v['v']
                     advise=v['advisory']
@@ -53,30 +56,50 @@ def getSafetyDBOut(savepath):
                     depv.append(d.name)
                     depv.append(d.version)
                     depv.append(d.filepath)
+                    depv.append(d.bloated)
                     depv.append(cve)
                     depv.append(vers)
                     depv.append(advise)
                     deps.append(depv)
+            if(not d.vulnerable):
+                #when the vulnerable flag is still False after the checkVersions loop, it means that the dependency is in the db but that our analyzed version is not vulnerable
+                depv=[]
+                depv.append(d.name)
+                depv.append(d.version)
+                depv.append(d.filepath)
+                depv.append(d.bloated)
+                depv.append("no one")
+                depv.append("no one")
+                depv.append("no one")
+                deps.append(depv)
         except KeyError:
             #when a dependency is not vulnerable, is not in the db and try to use its name as a key, throws an exception
             depv=[]
             depv.append(d.name)
             depv.append(d.version)
             depv.append(d.filepath)
+            depv.append(d.bloated)
             depv.append("no one")
             depv.append("no one")
             depv.append("no one")
             deps.append(depv)
-            continue
     
     #create a pandas dataframe and saves it on a csv table file
     safetyDBFilter=pd.DataFrame(deps,columns=cols)
     safetyDBFilter.to_csv(savepath+"safetyDB_out.csv")
             
     file3.close()
+    
 
+def setBloated(depName,filePath):
+    for d in dependencies:
+        if(d.name==depName and d.filepath==filePath):
+            d.bloated=True
+            return True
 
-def getUnusedRequirements(pycd_savepath,propath,savepath):
+    return False
+
+def getUnusedRequirements(propath,savepath):
     #open the fawlatydeps_out.txt file in read/write mode
     file1=open(savepath+"fawltydeps_out.txt","w+")
     
@@ -85,12 +108,6 @@ def getUnusedRequirements(pycd_savepath,propath,savepath):
     
     #move file pointer to the beginnig of the file
     file1.seek(0)
-
-    #open the csv to read it with pandas
-    csv1=pd.read_csv(pycd_savepath)
-    
-    #open the file where the dependencies declaration are going to be written
-    #file2=open(savepath+"requirements-unused.txt", "w")
     
     #excluding the first two lines of the file
     file1.readline()
@@ -98,71 +115,77 @@ def getUnusedRequirements(pycd_savepath,propath,savepath):
     
     #read the first line with a dependency name
     liner=file1.readline()
-    
+
     while True:
         #delete from the line, whitespace characters and newlines
         line=liner.rstrip('\r\n')
         line=line.strip()
-        try:
+
         #when the line starts with a dash, is the one with the dependency name
+        try:
             if(line[0]=="-"):
                 dep=line[3:(len(line)-14)]#take the dependency name substring from the line
-                vers = csv1.loc[csv1['dep'] == dep, 'version'].values#take the dependency versions from pycd_out.csv
-                files = list(csv1.loc[csv1['dep'] == dep, 'filepath'].values)#take the filepaths where dependency is declared from pycd_out.csv
                 linep=file1.readline()
-                
-                #print(dep)
-                #print(files)
-            
-                #read all the lines after the one with the dependency name, until it finds a new one with another dependency name
+        
                 while(linep!="" and linep[0]!="-"):
                     #extract the filepath from the line
                     indS=linep.index('/')
                     filePath=linep[indS:len(linep)-1]
-                    try:
-                        ind=files.index(filePath)
-                    except ValueError as e:
-                        msg=str(e)
-                        path=msg[1:len(msg)-16]
-                        print("Pycd has some problems with config declaration in this file:")
-                        print(path)
-                        print("The results are affected by this issue, some dependencies are not considered")
-                
-                    #create a new string that contains the name and the version of the dependency
-                    depName=''.join([dep,vers[ind]])
-                    depName=depName.replace(' ','')
-                    depName+='\n'
-                
-                    #create a new Dependency object and appends it in dependencies list
-                    aDep=dependency.Dependency(dep,vers[ind],filePath)
-                    dependencies.append(aDep)
-                
+                    #check if the dependency is bloated
+                    if(not setBloated(dep,filePath)):
+                        print("PyCD is missing some configuration files, the output is incomplete. Execution interrupted")
+                        os.remove(savepath+"fawltydeps_out.txt")
+                        exit(0)
                     #read a new line from file1
                     linep=file1.readline()
-                
-            liner=linep
-            if(liner==""):
-                break
         except IndexError:
-                #if the fawltydep_out.txt has less than three lines, it means that the project has no bloated dependencies
-                file1.seek(0)
-                file1.readline()
-                msg=file1.readline()
-                if(msg=="No unused dependencies detected."):
-                    print("This project has no bloated dependencies")
-                else:
-                    print("Fawltydeps gives errors, you need to manually execute the tool to check the issues")
-                os.remove(savepath+"fawltydeps_out.txt")
-                exit(0)
-            
+            #if the fawltydep_out.txt has less than three lines, it means that the project has no bloated dependencies
+            file1.seek(0)
+            file1.readline()
+            msg=file1.readline()
+            msg=msg.rstrip('\r\n')
+            if(msg=="No unused dependencies detected."):
+                print("This project has no bloated dependencies")
+                return
+            else:
+                print("Fawltydeps returns errors and you need to manually execute the tool to check the issues, the output is incomplete. Execution interrupted")
+
+            os.remove(savepath+"fawltydeps_out.txt")
+            exit(0)
+
+                
+        liner=linep
+        if(liner==""):
+            break
+
     #closes the file streams            
     file1.close()
-    
-    os.remove(savepath+"fawltydeps_out.txt")
 
+    os.remove(savepath+"fawltydeps_out.txt")
 
 def getPyCDOut(propath,pycd_savepath):
     child=subprocess.run(["python","./GetDep_ast.py",propath,pycd_savepath])
+
+    try:
+        csv=pd.read_csv(pycd_savepath)
+        csvd=csv.drop_duplicates(subset=["dep","version","filepath"], keep="first")
+        csvdd=csvd.drop(columns=['Unnamed: 0'])
+    except FileNotFoundError:
+        print("PyCD has some issues scanning this project, pycd_out.csv file does not exists. Execution interruepted")
+        exit(0)
+
+    if(len(csvd)==0):
+        print("PyCD has some issues scanning this project, pycd_out.csv is an empty file. Execution interruepted")
+        exit(0)
+
+    #also save the pycd output in a list of Dependency objects
+    for rowInd in csvdd.iterrows():
+        row=rowInd[1]
+        aDep=dependency.Dependency(row['dep'],row['version'],row['filepath'])
+        dependencies.append(aDep)
+
+    os.remove(pycd_savepath)
+    csvdd.to_csv(pycd_savepath)
 
 
 def launch(propath,savepath):
@@ -178,16 +201,15 @@ def launch(propath,savepath):
     getPyCDOut(propath,pycd_savepath)
     print("PyCD output obtained in csv format (pycd_out.csv)")
     
-    #2. obtain the outupt of falwtydeps and creates a requirements file
-    getUnusedRequirements(pycd_savepath,propath,savepath)
-    #print("Unused requirements file obtained with FalwtyDeps")
+    #2. obtain the outupt of falwtydeps and check which dependencies are bloated
+    getUnusedRequirements(propath,savepath)
     
     #3. obtain a csv format of bloated&vulnerable dependencies, filtering safetyDB
     getSafetyDBOut(savepath)
     print("SafetyDB filtered output obtained in csv format (report.csv)")
     print()
     
-    getShellOutputJ(savepath)
+    getShellOutput(savepath)
     print()
     print("You can find all the detailed output in the files stored in " + savepath + " directory")
